Implement "undo last move" action

+++ next-items.h
@@ -8,6 +8,6 @@
 
 @interface NextItems : MUIArea
 
--(VOID) setNextItems: (LONG[3])nextItems;
+-(VOID) setNextItems: (BYTE[3])nextItems;
 
 @end
